Adding transactions, gas, types, and utils to api directory. Further migration of RPC methods from handler.go.
